Add files via upload
added requirements.txt so end user knows how to access the application. Game is not a standalone and needs insertion of values,creation of tables which has been given in the requirements.txt. Although the dataset given is pretty small, it still gives a clear idea on how the core concepts and sub-applications of the application work.